resolved pull request comments

+++ src/main/resources/db/migration/V3__create_sales_employee_table.sql
@@ -1,6 +1,8 @@
-CREATE TABLE salesEmployee (
-           salesID smallint PRIMARY KEY AUTO_INCREMENT,
-           commRate float(13,2),
+CREATE TABLE salesEmployee
+(
+    salesID  smallint PRIMARY KEY AUTO_INCREMENT,
+    commRate float(13, 2
+) ,
            employeeID smallint,
            FOREIGN KEY (employeeID) REFERENCES employee(employeeID)
 );
